Refactor: WebPageReader 구조 변경
